requirements.txt
urllib3==1.26.6
PyYAML
requests
pyTelegramBotAPI
pytest